Create README.md
The Viterbi decoding method uses the maximum likelihood decoding (MLD) algorithm, which finds the most likely pattern from the received data. In our project, using the Hidden Markov Model, Viterbi Decoding outperforms uncoded communication remarkably in most cases when Eb/N0 of system is not too bad. This feature becomes even more significant when the ratio of '0's is high in the input bitstream. This result shows that Viterbi Decoding is a decent choice for digital communication.